+++ sky-server/src/main/java/com/sky/mapper/SetmealDishMapper.java
@@ -7,5 +7,5 @@ import java.util.List;
 @Mapper
 public interface SetmealDishMapper {
     //根据菜品id集合查询套餐id集合
-    List<Integer> getSetmealIdsByDishIds(List<Integer> dishIds);
+    List<Integer> getSetmealIdsByDishIds(List<Long> dishIds);
 }
